% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inloop.R
\name{inloop}
\alias{inloop}
\title{inloop}
\usage{
inloop(xset, yset, xvar, yvar)
}
\arguments{
\item{xset}{The first dataframe}

\item{yset}{The reference dataframe}

\item{xvar}{Variable holding values to check for in the second dataframe. This should be enclosed in quotation marks.}

\item{yvar}{Reference dataframe variable that is is checked against. This should be enclosed in quotation marks}
}
\description{
Loop that checks to see if a value of a variable in one set is in another set. Returns a vector of the same length as the xset, 
which in practice can be added to the xset as an additional column.
}
\examples{

# Unfortunately - I haven't created a dataset yet to demo this. However, this is the loop in practice.
female.df$in_hhmem <- NA
female.df$in_hhmem <- inloop(female.df, hh_mem.df, "clus_FemaleID", "clus_FemaleID")
female.df$in_hhmem <- ifelse(female.df$in_hhmem == 1, 0, 1)

}
\author{
J.W. Rozelle
}
