% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/killDbConnections.R
\name{killDbConnections}
\alias{killDbConnections}
\title{killDbConnections}
\usage{
killDbConnections()
}
\description{
Kills database connections
}
\examples{

# Quite simple - just use without parameters
killDbConnections()


}
\author{
Unknown
}
