% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odkbc_CheckAndDL.R
\name{odkbc_CheckAndDL}
\alias{odkbc_CheckAndDL}
\title{odkbc_CheckAndDL}
\usage{
odkbc_CheckAndDL()
}
\description{
This function is NOT meant to be called directly. It is embedded within other ODK functions to make sure that ODK is downloaded in the correct directory
}
\examples{

# Run function
odkbc_CheckAndDL()


}
\author{
J.W. Rozelle
}
