% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procFreq.R
\name{procFreq}
\alias{procFreq}
\title{procFreq}
\usage{
procFreq(data, vari_col, indic, group_col)
}
\arguments{
\item{data}{The dataframe that contains the information you want}

\item{vari_col}{The the variable column. This should be enclosed in quotations marks. For example, data$variable would be noted here as "variable"}

\item{indic}{Optional parameter for the name of the indicator}

\item{group_col}{Optional parameter.This would be the column used to stratify results.}
}
\description{
This function was created because the available options in R for frequency tables of 1 / 0 indicators was insufficient.
}
\examples{

# Get the mtcars data
data(mtcars)

# Run the simplest version of the variable
procFreq(mtcars, "am")


# Name the indicator and disaggregate by the cyl column
procFreq(data = mtcars, vari_col = "am", indic = "Indicator Description by cyl", group_col = "cyl")


}
\author{
J.W. Rozelle
}
