% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullForm.R
\name{pullForm}
\alias{pullForm}
\title{pullForm}
\usage{
pullForm(
  url = NULL,
  userid = NULL,
  passwd = NULL,
  formid = NULL,
  storage = getwd(),
  central = FALSE,
  project_id = NULL
)
}
\arguments{
\item{url}{The URL to the ODK Aggregate server. This may be something like `https://kc.humanitarianresponse.info/<USERID>/`. Be sure to include your userid here if the server requires it.}

\item{userid}{The user id for your odk aggregate server account. This must have permission to view submissions.}

\item{passwd}{The Password for your account. Be careful about keeping this in your R code. Best practice is likely to put this in your .Renviron file.}

\item{formid}{The unique formid string for your server.}

\item{storage}{Optional parameter, and defaults to your current working directory.}

\item{central}{Specify as TRUE if pulling from an ODK Central server. It is FALSE by default, which means that the command will assume an aggregate type server.}

\item{project_id}{When pulling from a central server, it is essential to specify a project_id.}
}
\description{
This function pulls forms from an odk server using ODK Briefcase. It's likely useful to set some of the parameters in variables, as they will be necessary to export the forms.
}
\note{
This function will check for and download ODK Briefcase. Make sure your firewalls allow for this the first time you run an odk function.
}
\examples{

# Set parameters
odkUserID <- "<USERNAME_GOES_HERE>"
odkURL <- paste0("https://kc.humanitarianresponse.info/", odkUserID, "/")
odkPW <- "<SECRET_PASSWORD>"
odkFormid <- "my_odk_form_id"
odkStorage <- paste0(my_directory, "/", "ODK")


# Run command
pullForm(odkURL, odkUserID, odkPW, odkFormid, odkStorage)


}
\author{
J.W. Rozelle
}
