% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stata_svyprop.R
\name{stata_svyprop}
\alias{stata_svyprop}
\title{stata_svyprop}
\usage{
stata_svyprop(design, indicator, level = 0.95)
}
\arguments{
\item{design}{survey design object}

\item{indicator}{The indicator for used to calculate. Should be in binary format (i.e. 0 and 1)}

\item{level}{Optional parameter the confidence level, defaults to 0..95}
}
\description{
This function was intended to simplify coding for survey data. This uses the 'survey' package, and inputs necessary for a similar calculation for proportion CI's as STATA. Additionally, it outputs the data in a dataframe format for viewing
}
\examples{





}
\author{
J.W. Rozelle
}
