% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waveSampling.R
\name{waveSampling}
\alias{waveSampling}
\title{Show a wave being sampled}
\usage{
waveSampling(wave, plot_every)
}
\arguments{
\item{wave}{A Wave object}

\item{plot_every}{Sample rate to show}
}
\description{
Show a wave being sampled
}
