% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportForm.R
\name{exportForm}
\alias{exportForm}
\title{exportForm}
\usage{
exportForm(
  formid = NULL,
  exportDir = NULL,
  fileName = NULL,
  storage = getwd(),
  pemKey = NULL,
  geojson = FALSE,
  mediaInclude = TRUE,
  splitMultiple = TRUE,
  pullBefore = FALSE,
  rm_group_names = TRUE,
  oc = TRUE
)
}
\arguments{
\item{formid}{The unique formid string for your server.}

\item{exportDir}{Export Directory. If left NULL, the function will use odk_export in the working directory and create odk_export in the working directory if it doesn't exist.}

\item{fileName}{Name of csv, must end in '.csv'}

\item{storage}{Optional parameter, and defaults to your current working directory.}

\item{pemKey}{Optional parameter, if the data is encrypted - this should be the filepath to the PEM key}

\item{geojson}{Optional, logical parameter, defaults to TRUE. Mark false if you do not wish to include a geojson with geodata}

\item{mediaInclude}{Optional, logical parameter, defauts to TRUE. If you wish to exclude media, mark FALSE}

\item{splitMultiple}{Optional, logical parameter, defaults to TRUE. IF you do not wish to split 'select_multiple' type questions into separate columns, mark false.}

\item{pullBefore}{Logical parameter, defaults to False. If you want to pull before export. Regardless, at least one manual pull is required fefore .}

\item{rm_group_names}{Optional, defualts to TRUE. If you wish to include group names, mark FALSE}

\item{oc}{Optional, defaults to TRUE. Set to true (default behavior), this will overwrite a previously exported CSV. Set to false, it will not. Note that you must change the name of your csv for the function to work if set to FALSE and you've already exported to the same directory with the same filename}
}
\description{
This function pulls forms from an odk server using ODK Briefcase. It's likely useful to set some of the parameters in variables, as they will be necessary to export the forms.
}
\note{
This function will check for and download ODK Briefcase. Make sure your firewalls allow for this the first time you run an odk function.
}
\examples{

# specify arguments

odkbc_storage   <- "<your/odk/storage/path>"
odkbc_export    <- "<your/export/file/path>"
random_id       <- "random_form_id"
random_filename <- "/exportdataname.csv"
odk_username    <- "MY_USERNAME"
odk_url         <- paste0("https://kc.humanitarianresponse.info/", odk_username)

pullForm(odk_url, odk_username, "PASSWORD", random_id, storage = odkbc_storage)

# Export ODK
exportForm(random_id, odkbc_export, random_filename, odkbc_storage)



}
\author{
J.W. Rozelle
}
