% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odkbc_CheckAndDL.R
\name{odkbc_CheckAndDL}
\alias{odkbc_CheckAndDL}
\title{odkbc_CheckAndDL}
\usage{
odkbc_CheckAndDL()
}
\description{
This function is embedded within other ODK functions to make sure that ODK is downloaded in the correct directory
}
\note{
This function is not necessary to by itself. Anytime ODK is needed in this package, the function is run.
}
\examples{

# Run function
odkbc_CheckAndDL()


}
\author{
J.W. Rozelle
}
